% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asym.env.R
\name{asym.env}
\alias{asym.env}
\title{Asyptotic covariance of vec(beta) for a fitted envelope model.}
\usage{
asym.env(object, ...)
}
\arguments{
\item{object}{A fitted env object.}
}
\value{
\item{covMatrix}{The asymptotic covariance of vec(beta). An rp by rp matrix. The covariance matrix returned is asymptotic. For the actual standard errors, multiply by 1/n.}
\item{asySE}{The asymptotic standard error for elements in beta under the envelope model. An r by p matrix. The standard errors returned are asymptotic, for actual standard errors, multiply by 1/sqrt(n).}
\item{ratio}{The asymptotic standard error ratio of the standard multivariate linear regression estimator over the envelope estimator, for each element in beta. An r by p matrix.}
}
\description{
This function computes the asymptotic covariance of vec(beta) and asymptotic standard error for elements in beta for a fitted envelope model.
}
\references{
Cook, R. Dennis, Bing Li, and Francesca Chiaromonte. "Envelope models for parsimonious and efficient multivariate linear regression." \emph{Statist. Sinica} 20 (2010): 927-1010.
}
\seealso{
\code{choose_env} for choosing the dimension of the envelope subspace.
}
