% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asym.spenv.R
\name{asym.spenv}
\alias{asym.spenv}
\title{Asymptotic covariance of vec(beta) for a fitted sparse envelope model}
\usage{
asym.spenv(object)
}
\arguments{
\item{object}{A fitted spenv object.}
}
\value{
\item{covMatrix}{The asymptotic covariance of vec(beta). An rp by rp matrix. The covariance matrix returned is asymptotic. For the actual standard errors, multiply by 1/n.}
\item{asySE}{The asymptotic standard error for elements in beta under the envelope model. An r by p matrix. The standard errors returned are asymptotic, for actual standard errors, multiply by 1/sqrt(n).}
\item{ratio}{The asymptotic standard error ratio of the standard multivariate linear regression estimator over the envelope estimator, for each element in beta. An r by p matrix.}
}
\description{
Asymptotic covariance of vec(beta) for a fitted sparse envelope model
}
\references{
Su Z, Zhu G, Chen X, Yang Y. Sparse envelope model: efficient estimation and response variable selection in multivariate linear regression. Biometrika. 2016 Sep 1;103(3):579-93.
}
\seealso{
\code{choose_spenv} for choosing the dimension of envelope subspace.
}
\author{
Guangyu Zhu \href{mailto:guangyuzhu@uri.edu}{guangyuzhu@uri.edu}
}
