% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Berkeley}
\alias{Berkeley}
\title{Berkeley guidance data}
\format{
\describe{
\item{X}{X=1 for boy, and X=0 for girl.}
\item{Y}{Height measurements at 31 ages.}
}
}
\usage{
data(Berkeley)
}
\description{
Berkeley guidance data (Tuddenham and Snyder, 1953) includes height measurements for 39 boys and 54 girls born in 1928-1929 in Berkeley, CA.
}
\references{
Tuddenham, R. D. and M. M. Snyder (1953). Physical growth of California boys and girls from birth
to eighteen years. Publications in child development. University of California, Berkeley 1(2), 183–364.
}
\keyword{datasets}
