% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_spenv.R
\name{choose_spenv}
\alias{choose_spenv}
\title{choose_spenv}
\usage{
choose_spenv(X, Y)
}
\arguments{
\item{X}{Predictors. An n by p matrix, p is the number of predictors. The predictors can be univariate or multivariate, discrete or continuous.}

\item{Y}{Multivariate responses. An n by r matrix, r is the number of responses and n is the number of observations. The responses must be continuous variables.}

\item{dims}{A vector. The dimensions to be chosen from. The default is 0 to r.}

\item{maxiter}{Maximum number of iterations. Default value: 100.}

\item{ftol}{Tolerance parameter for F. Default value: 1e-2.}

\item{verbose}{Flag for print out model fitting process, logical 0 or 1. Default value: 0.}
}
\value{
\item{result}{Dimensions of the envelope subspace chosen by BIC, AIC, and LRT, and the corresponding tuning parameter chosen for the dimension.}
\item{detail}{-2*Loglik, BIC, AIC, degrees of freedom, and p-values for LRT test.}
}
\description{
Select the dimension of the envelope subspace and the tuning parameter using Bayesian information
criterion, Akaike information criterion, and Likelihood ratio testing.
}
\details{
Choose u and lambda for sparse envelope model.
}
\references{
The codes are implemented based on the algorithm in Z. Su, G. Zhu, and X. Chen (2015).

Su Z, Zhu G, Chen X, Yang Y. Sparse envelope model: efficient estimation and response variable selection in multivariate linear regression. Biometrika. 2016 Sep 1;103(3):579-93.
}
