% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_spxenv.R
\name{choose_spxenv}
\alias{choose_spxenv}
\title{Choose u and lambda for envelope-based sparse partial least squares model.}
\usage{
choose_spxenv(X, Y)
}
\arguments{
\item{X}{Predictors. An n by p matrix, p is the number of predictors. The predictors can be univariate or multivariate, discrete or continuous.}

\item{Y}{Multivariate responses. An n by r matrix, r is the number of responses and n is the number of observations. The responses must be continuous variables.}

\item{dims}{A vector. The dimensions to be chosen from. The default is 0 to p.}

\item{maxiter}{Maximum number of iterations. Default value: 100.}

\item{ftol}{Tolerance parameter for F. Default value: 1e-2.}

\item{verbose}{Flag for print out model fitting process, logical 0 or 1. Default value: 0.}
}
\value{
\item{result}{Dimensions of the envelope subspace chosen by BIC, AIC, and LRT, and the corresponding tuning parameter chosen for the dimension.}

\item{detail}{-2*Loglik, BIC, AIC, degrees of freedom, and p-values for LRT test.}
}
\description{
Select the dimension of the envelope subspace and the tuning parameter using Bayesian information
criterion, Akaike information criterion, and Likelihood ratio testing.
}
\references{
The codes are implemented based on the algorithm in Guangyu Zhu. Zhihua Su. "Envelope-based sparse partial least squares." Ann. Statist. 48 (1) 161 - 182, February 2020. \href{https://doi.org/10.1214/18-AOS1796}
}
\author{
Guangyu Zhu and Zhihua Su
}
