% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mussel}
\alias{mussel}
\title{The New Zealand mussels data}
\format{
The data frame contains the following components:

\describe{
\item{X}{X=0 if it was found to have pea crabs living within the shell, X=1 if it was free of pea crabs.}
\item{Y}{Logarithms of shell height, shell width, shell length, shell mass, mussel mass, viscera mass.}
}
}
\usage{
data(mussel)
}
\description{
The New Zealand mussels data.
}
\details{
The New Zealand mussels data.
}
\references{
Cook, R. Dennis, Bing Li, and Francesca Chiaromonte. "Envelope models for parsimonious and efficient multivariate linear regression." Statist. Sinica 20 (2010): 927-1010.
}
\keyword{datasets}
