% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SAT}
\alias{SAT}
\title{SAT score data}
\format{
The data frame contains the following components:

\describe{
\item{sat}{SAT score;}
\item{takers}{percentage of the total eligible students in the state who took the exam;}
\item{income}{the median income of families of the test takers;}
\item{years}{the average number of years that the test takers had formal studies in social sciences, natural sciences, and humanities;}
\item{public}{the percentage of the test takers who attended public secondary schools;}
\item{expend}{the total state expenditure on secondary schools;}
\item{rank}{the median percentile ranking of the test takers within their secondary school classes.}
}
}
\source{
Ramsey, F. and Schafer, D. (2012). The Statistical Sleuth: A Course in Methods of Data Analysis. Boston: Cengage Learning.
}
\usage{
data(SAT)
}
\description{
The average SAT score of the fifty states in the U.S. in 1982, as well as six variables that are used to predict the average SAT score.
}
\details{
SAT score data.
}
\examples{
data(SAT)
SAT = as.data.frame(scale(SAT, scale = FALSE))
new = model.matrix(sat ~ (takers + income + years + public + expend + rank)^2, SAT)
X = as.matrix(new[, c(2:22)])
Y = as.matrix(SAT[, 1])
m1 = spxenv(X, Y, u = 3)

}
\keyword{datasets}
