% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wheatprotein}
\alias{wheatprotein}
\title{Wheat Protein Data}
\format{
A data frame with 50 observations on the following 8 variables:

\describe{
\item{\code{V1} to \code{V6}}{Measurements of the reflectance of NIR radiation by the wheat samples at 6 wavelengths in the range 1680-2310 nm. The measurements were made on the log(1/reflectance) scale.}
\item{\code{V7}}{The protein content of each sample (in percent).}
\item{\code{V8}}{Binary indicator, 0 for high protein content and 1 for low protein content. The cutoff point is if the protein content is smaller than 9.75.}
}
}
\usage{
data(wheatprotein)
}
\description{
The protein content of ground wheat samples.
}
\details{
Wheat Protein Data.

The data are the result of an experiment to calibrate a near
infrared reflectance (NIR) instrument for measuring the protein
content of ground wheat samples. The protein content of each
sample (in percent) was measured by the standard Kjeldahl method.
In Fearn (1983), the problem is to find a linear combination of
the measurements that predicts protein content. The estimated
coefficients can then be entered into the instrument allowing the
protein content of future samples to be read directly. The first
24 cases were used for calibration, and the last 26 samples were
used for prediction.
}
\examples{
data(wheatprotein)
X <- wheatprotein[, 8]
Y <- wheatprotein[, 1:6]
choose_env(X, Y)
m1 = env(X, Y, 1)
m1$beta
}
\references{
Fearn, T. (1983). A misuse of ridge regression in the calibration of a near infrared reflectance instrument.
}
\keyword{datasets}
